refactor: move resolvers to `View`
