+++ packages/laravel/tests/Laravel/PropertiesResolver/RequestPropertiesResolverTest.php
@@ -3,7 +3,7 @@
 use Hybridly\Support\CaseConverter;
 use Hybridly\Support\Deferred;
 use Hybridly\Support\Partial;
-use Hybridly\Support\PropertiesResolver;
+use Hybridly\View\PropertiesResolver;
 use Hybridly\View\Factory;
 use Illuminate\Contracts\Support\Arrayable;
